requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# API
gcsfs
google-cloud-storage
mlflow
s3fs
requests
#openalexapi-0.0.1a0

# get data
pdfminer.six==20220524

# data preprocessing
numpy
pandas
bs4
regex
nltk
scikit-learn
gensim==4.2.0

# model
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0

# tests/linter
black
coverage
flake8
pytest
yapf

# utilities
six>=1.14
joblib
memoized-property
termcolor
colorama

# api
fastapi
uvicorn
